# Conflicts:
#	backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/endpoint/SpotEndpoint.java
#	backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/endpoint/exceptionhandler/GlobalExceptionHandler.java
#	backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/service/SpotService.java
#	backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/service/impl/SimpleSpotService.java

+++ backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/service/impl/SimpleSpotService.java
@@ -1,5 +1,6 @@
 package at.ac.tuwien.sepm.groupphase.backend.service.impl;
 import at.ac.tuwien.sepm.groupphase.backend.entity.Spot;
+import at.ac.tuwien.sepm.groupphase.backend.repository.LocationRepository;
 import at.ac.tuwien.sepm.groupphase.backend.exception.NotFoundInDatabaseException;
 import at.ac.tuwien.sepm.groupphase.backend.exception.ServiceException;
 import at.ac.tuwien.sepm.groupphase.backend.exception.ValidationException;
@@ -8,10 +9,12 @@ import at.ac.tuwien.sepm.groupphase.backend.repository.LocationRepository;
 import at.ac.tuwien.sepm.groupphase.backend.repository.SpotRepository;
 import at.ac.tuwien.sepm.groupphase.backend.service.LocationService;
 import at.ac.tuwien.sepm.groupphase.backend.service.SpotService;
+import org.hibernate.service.spi.ServiceException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
+
+import javax.xml.bind.ValidationException;
 import java.lang.invoke.MethodHandles;
 
 @Service
@@ -54,14 +57,17 @@ public class SimpleSpotService implements SpotService {
         }
         return spotRepository.save(spot);
     }
-//ToDo: Fix notFoundinDatabaseException
+
     @Override
-    public void deleteById(Long id) throws NotFoundInDatabaseException {
+    public void deleteById(Long id) throws ValidationException {
         LOGGER.debug("Delete Spot with id {}", id);
-        try {
-            spotRepository.deleteById(id);
-        }catch (EmptyResultDataAccessException e){
-            throw new NotFoundInDatabaseException("Spot does not exist");
+        var spot = spotRepository.findById(id);
+        if(spot.isEmpty()){
+            throw new ValidationException("Spot does not exist");
+        }
+        spotRepository.deleteById(id);
+        if (spotRepository.findLocationWithSpot(spot.get().getLocation().getId()).isEmpty()){
+            locationRepository.deleteById(spot.get().getLocation().getId());
         }
     }
 }
